+++ test/test_bench_examples.py
@@ -30,8 +30,8 @@ from bencher.example.example_dataframe import example_dataset
 from bencher.example.optuna.example_optuna import optuna_rastrigin
 from bencher.example.example_sample_cache import example_sample_cache
 from bencher.example.example_strings import example_strings
-from bencher.example.example_yaml_sweep import example_yaml_sweep
-from bencher.example.example_yaml_dict_sweep import example_yaml_dict_sweep
+from bencher.example.example_yaml_sweep_list import example_yaml_sweep
+from bencher.example.example_yaml_sweep_dict import example_yaml_dict_sweep
 from bencher.example.example_image import example_image, example_image_vid
 from bencher.example.example_video import example_video
 from bencher.example.example_filepath import example_filepath
